+++ griptape/drivers/prompt/base_prompt_driver.py
@@ -1,14 +1,22 @@
 from __future__ import annotations
+
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Optional
 from collections.abc import Iterator
-from attrs import define, field, Factory
-from griptape.events import StartPromptEvent, FinishPromptEvent, CompletionChunkEvent
-from griptape.mixins.serializable_mixin import SerializableMixin
-from griptape.utils import PromptStack
-from griptape.mixins import ExponentialBackoffMixin
+from typing import TYPE_CHECKING, Optional
+
+from attrs import Factory, define, field
+
+from griptape.common import (
+    BaseDeltaMessageContent,
+    DeltaMessage,
+    TextDeltaMessageContent,
+    PromptStack,
+    Message,
+    TextMessageContent,
+)
+from griptape.events import CompletionChunkEvent, FinishPromptEvent, StartPromptEvent
+from griptape.mixins import ExponentialBackoffMixin, SerializableMixin
 from griptape.tokenizers import BaseTokenizer
-from griptape.artifacts import TextArtifact
 
 if TYPE_CHECKING:
     from griptape.structures import Structure
@@ -16,7 +24,7 @@ if TYPE_CHECKING:
 
 @define(kw_only=True)
 class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
-    """Base class for Prompt Drivers.
+    """Base class for the Prompt Drivers.
 
     Attributes:
         temperature: The temperature to use for the completion.
@@ -39,38 +47,28 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
 
     def before_run(self, prompt_stack: PromptStack) -> None:
         if self.structure:
-            self.structure.publish_event(
-                StartPromptEvent(
-                    model=self.model,
-                    token_count=self.tokenizer.count_tokens(self.prompt_stack_to_string(prompt_stack)),
-                    prompt_stack=prompt_stack,
-                    prompt=self.prompt_stack_to_string(prompt_stack),
-                )
-            )
+            self.structure.publish_event(StartPromptEvent(model=self.model, prompt_stack=prompt_stack))
 
-    def after_run(self, result: TextArtifact) -> None:
+    def after_run(self, result: Message) -> None:
         if self.structure:
             self.structure.publish_event(
                 FinishPromptEvent(
-                    model=self.model, result=result.value, token_count=self.tokenizer.count_tokens(result.value)
+                    model=self.model,
+                    result=result.value,
+                    input_token_count=result.usage.input_tokens,
+                    output_token_count=result.usage.output_tokens,
                 )
             )
 
-    def run(self, prompt_stack: PromptStack) -> TextArtifact:
+    def run(self, prompt_stack: PromptStack) -> Message:
         for attempt in self.retrying():
             with attempt:
                 self.before_run(prompt_stack)
 
                 if self.stream:
-                    tokens = []
-                    completion_chunks = self.try_stream(prompt_stack)
-                    for chunk in completion_chunks:
-                        self.structure.publish_event(CompletionChunkEvent(token=chunk.value))
-                        tokens.append(chunk.value)
-                    result = TextArtifact(value="".join(tokens).strip())
+                    result = self.__process_stream(prompt_stack)
                 else:
-                    result = self.try_run(prompt_stack)
-                    result.value = result.value.strip()
+                    result = self.__process_run(prompt_stack)
 
                 self.after_run(result)
 
@@ -90,20 +88,59 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
         """
         prompt_lines = []
 
-        for i in prompt_stack.inputs:
+        for i in prompt_stack.messages:
+            content = i.to_text()
             if i.is_user():
-                prompt_lines.append(f"User: {i.content}")
+                prompt_lines.append(f"User: {content}")
             elif i.is_assistant():
-                prompt_lines.append(f"Assistant: {i.content}")
+                prompt_lines.append(f"Assistant: {content}")
             else:
-                prompt_lines.append(i.content)
+                prompt_lines.append(content)
 
         prompt_lines.append("Assistant:")
 
         return "\n\n".join(prompt_lines)
 
     @abstractmethod
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact: ...
+    def try_run(self, prompt_stack: PromptStack) -> Message: ...
 
     @abstractmethod
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]: ...
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]: ...
+
+    def __process_run(self, prompt_stack: PromptStack) -> Message:
+        result = self.try_run(prompt_stack)
+
+        return result
+
+    def __process_stream(self, prompt_stack: PromptStack) -> Message:
+        delta_contents: dict[int, list[BaseDeltaMessageContent]] = {}
+        usage = DeltaMessage.Usage()
+
+        # Aggregate all content deltas from the stream
+        deltas = self.try_stream(prompt_stack)
+        for delta in deltas:
+            usage += delta.usage
+
+            if delta.content is not None:
+                if delta.content.index in delta_contents:
+                    delta_contents[delta.content.index].append(delta.content)
+                else:
+                    delta_contents[delta.content.index] = [delta.content]
+
+            if isinstance(delta.content, TextDeltaMessageContent):
+                self.structure.publish_event(CompletionChunkEvent(token=delta.content.text))
+
+        # Build a complete content from the content deltas
+        content = []
+        for deltas in delta_contents.values():
+            text_deltas = [delta for delta in deltas if isinstance(delta, TextDeltaMessageContent)]
+            if text_deltas:
+                content.append(TextMessageContent.from_deltas(text_deltas))
+
+        result = Message(
+            content=content,
+            role=Message.ASSISTANT_ROLE,
+            usage=Message.Usage(input_tokens=usage.input_tokens, output_tokens=usage.output_tokens),
+        )
+
+        return result