+++ griptape/drivers/prompt/cohere_prompt_driver.py
@@ -4,8 +4,17 @@ from collections.abc import Iterator
 from attrs import define, field, Factory
 from griptape.artifacts import TextArtifact
 from griptape.drivers import BasePromptDriver
-from griptape.utils import PromptStack, import_optional_dependency
-from griptape.tokenizers import BaseTokenizer, CohereTokenizer
+from griptape.tokenizers import CohereTokenizer
+from griptape.common import (
+    PromptStack,
+    Message,
+    DeltaMessage,
+    TextMessageContent,
+    BaseMessageContent,
+    TextDeltaMessageContent,
+)
+from griptape.utils import import_optional_dependency
+from griptape.tokenizers import BaseTokenizer
 
 if TYPE_CHECKING:
     from cohere import Client
@@ -29,41 +38,71 @@ class CoherePromptDriver(BasePromptDriver):
         default=Factory(lambda self: CohereTokenizer(model=self.model, client=self.client), takes_self=True)
     )
 
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
+    def try_run(self, prompt_stack: PromptStack) -> Message:
         result = self.client.chat(**self._base_params(prompt_stack))
+        usage = result.meta.tokens
 
-        return TextArtifact(value=result.text)
+        return Message(
+            content=[TextMessageContent(TextArtifact(result.text))],
+            role=Message.ASSISTANT_ROLE,
+            usage=Message.Usage(input_tokens=usage.input_tokens, output_tokens=usage.output_tokens),
+        )
 
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
         result = self.client.chat_stream(**self._base_params(prompt_stack))
 
         for event in result:
             if event.event_type == "text-generation":
-                yield TextArtifact(value=event.text)
+                yield DeltaMessage(content=TextDeltaMessageContent(event.text, index=0))
+            elif event.event_type == "stream-end":
+                usage = event.response.meta.tokens
 
-    def _prompt_stack_input_to_message(self, prompt_input: PromptStack.Input) -> dict:
-        if prompt_input.is_system():
-            return {"role": "SYSTEM", "text": prompt_input.content}
-        elif prompt_input.is_user():
-            return {"role": "USER", "text": prompt_input.content}
-        else:
-            return {"role": "ASSISTANT", "text": prompt_input.content}
+                yield DeltaMessage(
+                    usage=DeltaMessage.Usage(input_tokens=usage.input_tokens, output_tokens=usage.output_tokens)
+                )
 
-    def _base_params(self, prompt_stack: PromptStack) -> dict:
-        user_message = prompt_stack.inputs[-1].content
-
-        history_messages = [
-            self._prompt_stack_input_to_message(input) for input in prompt_stack.inputs[:-1] if input.content
+    def _prompt_stack_messages_to_messages(self, messages: list[Message]) -> list[dict]:
+        return [
+            {
+                "role": self.__to_role(message),
+                "content": [self.__prompt_stack_content_message_content(content) for content in message.content],
+            }
+            for message in messages
         ]
 
-        params = {
+    def _base_params(self, prompt_stack: PromptStack) -> dict:
+        last_input = prompt_stack.messages[-1]
+        user_message = last_input.to_text()
+
+        history_messages = self._prompt_stack_messages_to_messages(
+            [message for message in prompt_stack.messages[:-1] if not message.is_system()]
+        )
+
+        system_messages = prompt_stack.system_messages
+        if system_messages:
+            preamble = system_messages[0].to_text()
+        else:
+            preamble = None
+
+        return {
             "message": user_message,
+            "chat_history": history_messages,
             "temperature": self.temperature,
             "stop_sequences": self.tokenizer.stop_sequences,
             "max_tokens": self.max_tokens,
+            **({"preamble": preamble} if preamble else {}),
         }
 
-        if history_messages:
-            params["chat_history"] = history_messages
+    def __prompt_stack_content_message_content(self, content: BaseMessageContent) -> dict:
+        if isinstance(content, TextMessageContent):
+            return {"text": content.artifact.to_text()}
+        else:
+            raise ValueError(f"Unsupported content type: {type(content)}")
 
-        return params
+    def __to_role(self, message: Message) -> str:
+        if message.is_system():
+            return "SYSTEM"
+        elif message.is_user():
+            return "USER"
+        else:
+            return "CHATBOT"