+++ griptape/drivers/prompt/ollama_prompt_driver.py
@@ -5,8 +5,11 @@ from attrs import define, field, Factory
 from griptape.artifacts import TextArtifact
 from griptape.drivers import BasePromptDriver
 from griptape.tokenizers.base_tokenizer import BaseTokenizer
-from griptape.utils import PromptStack, import_optional_dependency
+from griptape.common import PromptStack, TextMessageContent
+from griptape.utils import import_optional_dependency
 from griptape.tokenizers import SimpleTokenizer
+from griptape.common import Message, DeltaMessage, TextDeltaMessageContent
+from griptape.common import ImageMessageContent
 
 if TYPE_CHECKING:
     from ollama import Client
@@ -46,24 +49,47 @@ class OllamaPromptDriver(BasePromptDriver):
         kw_only=True,
     )
 
-    def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
+    def try_run(self, prompt_stack: PromptStack) -> Message:
         response = self.client.chat(**self._base_params(prompt_stack))
 
         if isinstance(response, dict):
-            return TextArtifact(value=response["message"]["content"])
+            return Message(
+                content=[TextMessageContent(TextArtifact(value=response["message"]["content"]))],
+                role=Message.ASSISTANT_ROLE,
+            )
         else:
             raise Exception("invalid model response")
 
-    def try_stream(self, prompt_stack: PromptStack) -> Iterator[TextArtifact]:
+    def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
         stream = self.client.chat(**self._base_params(prompt_stack), stream=True)
 
         if isinstance(stream, Iterator):
             for chunk in stream:
-                yield TextArtifact(value=chunk["message"]["content"])
+                yield DeltaMessage(content=TextDeltaMessageContent(chunk["message"]["content"]))
         else:
             raise Exception("invalid model response")
 
     def _base_params(self, prompt_stack: PromptStack) -> dict:
-        messages = [{"role": input.role, "content": input.content} for input in prompt_stack.inputs]
+        messages = self._prompt_stack_to_messages(prompt_stack)
 
         return {"messages": messages, "model": self.model, "options": self.options}
+
+    def _prompt_stack_to_messages(self, prompt_stack: PromptStack) -> list[dict]:
+        return [
+            {
+                "role": message.role,
+                "content": message.to_text(),
+                **(
+                    {
+                        "images": [
+                            content.artifact.base64
+                            for content in message.content
+                            if isinstance(content, ImageMessageContent)
+                        ]
+                    }
+                    if any(isinstance(content, ImageMessageContent) for content in message.content)
+                    else {}
+                ),
+            }
+            for message in prompt_stack.messages
+        ]