+++ griptape/engines/rag/modules/base_rag_module.py
@@ -3,7 +3,7 @@ from concurrent import futures
 from typing import Callable, Any, Optional
 from attrs import define, field, Factory
 from griptape.engines.rag import RagContext
-from griptape.utils import PromptStack
+from griptape.common import PromptStack, Message
 
 
 @define(kw_only=True)
@@ -15,10 +15,7 @@ class BaseRagModule(ABC):
 
     def generate_query_prompt_stack(self, system_prompt: str, query: str) -> PromptStack:
         return PromptStack(
-            inputs=[
-                PromptStack.Input(system_prompt, role=PromptStack.SYSTEM_ROLE),
-                PromptStack.Input(query, role=PromptStack.USER_ROLE),
-            ]
+            messages=[Message(system_prompt, role=Message.SYSTEM_ROLE), Message(query, role=Message.USER_ROLE)]
         )
 
     def context_param(self, context: RagContext, key: str) -> Optional[Any]: