+++ griptape/engines/rag/modules/generation/prompt_generation_rag_module.py
@@ -1,5 +1,6 @@
 from typing import Callable
 from attrs import define, field, Factory
+from griptape.artifacts.text_artifact import TextArtifact
 from griptape.drivers import BasePromptDriver
 from griptape.engines.rag import RagContext
 from griptape.engines.rag.modules import BaseGenerationRagModule
@@ -40,7 +41,12 @@ class PromptGenerationRagModule(BaseGenerationRagModule):
 
                     break
 
-            context.output = self.prompt_driver.run(self.generate_query_prompt_stack(system_prompt, query))
+            output = self.prompt_driver.run(self.generate_query_prompt_stack(system_prompt, query)).to_artifact()
+
+            if isinstance(output, TextArtifact):
+                context.output = output
+            else:
+                raise ValueError("Prompt driver did not return a TextArtifact")
 
         return context