+++ griptape/utils/__init__.py
@@ -8,7 +8,6 @@ from .chat import Chat
 from .futures import execute_futures_dict
 from .futures import execute_futures_list
 from .token_counter import TokenCounter
-from .prompt_stack import PromptStack
 from .dict_utils import remove_null_values_in_dict_recursively, dict_merge
 from .file_utils import load_file, load_files
 from .hash import str_to_hash
@@ -38,7 +37,6 @@ __all__ = [
     "execute_futures_dict",
     "execute_futures_list",
     "TokenCounter",
-    "PromptStack",
     "remove_null_values_in_dict_recursively",
     "dict_merge",
     "Stream",